+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes
 - Fix crash when STDIN was interpreted as null (running windows version via wine)
 - Fix console window not opening when launching via Bootstrapper.exe
+- Fix formatting of redacted steam token (makes logs easier to read) 
 
 ## [0.7.1] - 2022-09-25